cleanup example, reorganize
